Create custompalette_bg_img
Copy and paste this into a new script in TamperMonkey to create a custom theme for your dashboard on the T blogging website. This version of the script lets you have a custom repeating background image too.. reclaim the tackiness of the old web!

The way this currently works is a bit janky (overwrites the 'navy' value with nothing so the background image behind it loads). Maybe I'll eventually figure out how to do it properly, maybe not.